refactor: remove shim layer
