--- src/obsidian-shim/development/column-events.ts
@@ -1,1 +1,0 @@
-export const useColumnEvents = () => {};
