--- src/obsidian-shim/development/export-events.ts
@@ -1,4 +1,0 @@
-import { TableState } from "src/shared/types";
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const useExportEvents = (_tableState: TableState) => {};
