--- src/obsidian-shim/development/menu-events.ts
@@ -1,6 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-export const useMenuEvents = (
-	_id: string,
-	_isOpen: boolean,
-	_isTextHighlighted: boolean
-) => {};
