--- src/obsidian-shim/development/notification.ts
@@ -1,5 +1,0 @@
-export class Notification {
-	constructor(message: string) {
-		console.log(message);
-	}
-}
