--- src/obsidian-shim/development/row-events.ts
@@ -1,1 +1,0 @@
-export const useRowEvents = () => {};
