--- src/obsidian-shim/development/vault-file.ts
@@ -1,14 +1,0 @@
-export class VaultFile {
-	extension: string;
-	basename: string;
-	path: string;
-	name: string;
-	modifiedTime: number;
-	parent: {
-		path: string;
-	} | null;
-}
-
-export const getVaultFiles = (): VaultFile[] => {
-	return [];
-};
